+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: passunca <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/09 12:18:48 by passunca          #+#    #+#             */
-/*   Updated: 2023/10/30 15:42:30 by passunca         ###   ########.fr       */
+/*   Updated: 2023/10/30 15:52:09 by passunca         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -78,7 +78,7 @@ t_list	*ft_lstmap(t_list *lst, void *(*f)(void *), void (*del)(void *));
 int		ft_putnchar_fd(char c, int fd, int n);
 int		ft_putstrn_fd(char *str, int fd, int n);
 int		ft_numlen(int n, int base);
-int		ft_uputnbr(unsigned int n);
+int		ft_uputnbr(size_t n);
 int		ft_putchar(char c);
 int		ft_putnbr(int n);
